+++ polaris-core/src/test/java/org/apache/polaris/core/entity/PolarisPrivilegeTest.java
@@ -131,7 +131,15 @@ public class PolarisPrivilegeTest {
         Arguments.of(100, PolarisPrivilege.TABLE_REMOVE_STATISTICS),
         Arguments.of(101, PolarisPrivilege.TABLE_REMOVE_PARTITION_SPECS),
         Arguments.of(102, PolarisPrivilege.TABLE_MANAGE_STRUCTURE),
-        Arguments.of(103, null));
+        Arguments.of(103, PolarisPrivilege.SEMANTIC_MODEL_LIST),
+        Arguments.of(104, PolarisPrivilege.SEMANTIC_MODEL_CREATE),
+        Arguments.of(105, PolarisPrivilege.SEMANTIC_MODEL_READ_PROPERTIES),
+        Arguments.of(106, PolarisPrivilege.SEMANTIC_MODEL_WRITE_PROPERTIES),
+        Arguments.of(107, PolarisPrivilege.SEMANTIC_MODEL_DROP),
+        Arguments.of(108, PolarisPrivilege.SEMANTIC_MODEL_FULL_METADATA),
+        Arguments.of(109, PolarisPrivilege.SEMANTIC_MODEL_LIST_GRANTS),
+        Arguments.of(110, PolarisPrivilege.SEMANTIC_MODEL_MANAGE_GRANTS_ON_SECURABLE),
+        Arguments.of(111, null));
   }
 
   @ParameterizedTest